requirements.txt
ipykernel
openai
python-dotenv
yfinance
pandas
matplotlib
colorama
streamlit
faker
SpeechRecognition>=3.10.0
playsound==1.2.2
pyaudio>=0.2.13